// peso_lucky_admin-main/src/pages/Record/Cancel/data.d.ts
export interface Params {}
export interface ListItem {
  actual_recycle_amount: number;
  affect_lines: number;
  cancel_at: string;
  cancel_id: number;
  create_at: string;
  game_name: string;
  seq: string;
  not_recycle_amount: number;
  operation_method: number;
  operator: string;
  order_cancel_total: number;
  recycle_win_amount: number;
  remark: string;
  return_lose_amount: number;
  user_cancel_total: number;
}